+++ lib/avo/test_helpers.rb
@@ -1,5 +1,7 @@
 module Avo
   module TestHelpers
+    include WaitForLoaded
+
     # Finds the wrapper element on the index view for the given field id and type, and associated with the given record id
     # Example usage:
     #   index_field_wrapper(id: "name", type: "text", record_id: 2)
